De-Bitu the PIC_ interfaces and downstream users
'Bitu' is an architecture-specific type that's sized large
enough to store a pointer, which on 32-bit systems is a 32-bit
type and similarly is a 64-bit type on 64-bit systems.

Bitu is perfect for storing memory addresses as integers, and
converting to an from emulated pointers to host-pointers.

Unfortunately, Bitu became heavily used throughout the the code
to store more than pointers and addresses.  It's used to hold
counters, data values, bit fields, and so on. This is despite
the fact that DOS isn't a 64-bit OS, and its data-types and
interfaces were never 64-bits in size. Likewise, devices that
are emulated also didn't use internal 64-bit
counters/bit-fields/ and so on.

Why do we care?

When compiling on a 32-bit platform (the default build target
for DOSBox upstream), this isn't much of a problem because
internally all of these Bitus become 32-bit types.  So passing
a Bitu into a function's explicit "uint32_t" argument is no
problem.

However, when building on 64-bit systems, all of these Bitu
variables balloon to 64-bits, and passing them into explicit
32-bit function arguments (or writen them into explicit 32-bit
variables) all of sudden is a problem (atleast for the
compiler).

Evidence of this can be seen in the Windows warnings counts:
The 272 warning on 32-bit swells to over 1,000 on 64-bit.

Isn't this risky?

Fortunately we already know that these Bitu values never
actually store content greater than 32-bit in size: proof of
this is the fact that DOSBox upstream's default build target
is Win32, and it's functioned well for decades.

So provided we don't touch the code that operates on memory
addresses, we already know that these general counters and data
types can be safely set to uint32_t (because that's actually
what they are anyway when building on Win32).

This commit is just a small step to steady fix this (this
commit only addresses roughly 50 of those Win64 warnings).
There are still thousands of instances left to fix in future
work.

This commit also makes use of PIC's floating point percentages
more consistent (as a portion was already based on float).

+++ src/cpu/core_full/load.h
@@ -515,14 +515,15 @@ l_M_Ed:
 		CPU_SW_Interrupt_NoIOPLCheck(1,GetIP());
 		continue;
 	case D_RDTSC: {
-		if (CPU_ArchitectureType<CPU_ARCHTYPE_PENTIUMSLOW) goto illegalopcode;
-		Bit64s tsc=(Bit64s)(PIC_FullIndex()*(double)(CPU_CycleAutoAdjust?70000:CPU_CycleMax));
-		reg_edx=(Bit32u)(tsc>>32);
-		reg_eax=(Bit32u)(tsc&0xffffffff);
+		if (CPU_ArchitectureType<CPU_ARCHTYPE_PENTIUMSLOW)
+			goto illegalopcode;
+		Bit64s tsc = (Bit64s)(PIC_FullIndex() * static_cast<float>(CPU_CycleAutoAdjust ? 70000 : CPU_CycleMax));
+		reg_edx = (Bit32u)(tsc >> 32);
+		reg_eax = (Bit32u)(tsc & 0xffffffff);
 		break;
-		}
+	}
 	default:
-		LOG(LOG_CPU,LOG_ERROR)("LOAD:Unhandled code %d opcode %X",inst.code.load,inst.entry);
+		LOG(LOG_CPU, LOG_ERROR)("LOAD:Unhandled code %d opcode %X", inst.code.load, inst.entry);
 		goto illegalopcode;
 }
